+++ src/euphorie/client/api/module.py
@@ -2,7 +2,7 @@ from five import grok
 from euphorie.json import export_image
 from ..navigation import FindPreviousQuestion
 from ..navigation import FindNextQuestion
-from ..module import ActionPlanView as BaseActionPlan
+from ..browser.module import ActionPlanView as BaseActionPlan
 from ..utils import HasText
 from ..model import Module
 from . import context_menu